tests/wpt/tests/infrastructure/testdriver/test_win_open_with_interaction.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>

  promise_test(async (t)=> new Promise((resolve, reject) => {
      let pointerMovementInOpenedWindow = false;
      let clickInOpenedWindow = false;
      let pointerMovementInTestWindow = false;
      let clickInTestWindow = false;
      window.pointerMoveInInteractedWindow = function() {
        pointerMovementInOpenedWindow = true;
      }
      window.clickInInteractedWindow = function() {
        clickInOpenedWindow = true;
        t.step(()=> {
          assert_true(pointerMovementInOpenedWindow);
          resolve();
        });
      }
      document.addEventListener("pointermove", function(event){
        pointerMovementInTestWindow = true;
      });
      document.addEventListener("click", function(event){
        clickInTestWindow = true;
        t.step(()=>  {
          assert_true(pointerMovementInTestWindow);
          reject("Interaction events fired in the test window!");
        });
      });

      let w = window.open("interacted_window.html", '_blank');
      t.add_cleanup(() => w.close());
    }), "Tests pointer move/click in a window opened with `window.open.`");
</script>
</body>
